+++ tests/test_configuration.py
@@ -45,6 +45,7 @@ class TestConstructor:
         iota = SplineProfile([2, 3, 4])
         surface = FourierRZToroidalSurface(NFP=2, sym=False)
         axis = FourierRZCurve([-0.2, 10, 0.3], [0.3, 0, -0.2], NFP=2, sym=False)
+        xsection = ZernikeRZToroidalSection(spectral_indexing="ansi")
         eq = Equilibrium(
             M=2,
             pressure=pressure,
@@ -64,17 +65,15 @@ class TestConstructor:
         np.testing.assert_allclose(axis.R_n, eq.Ra_n)
         np.testing.assert_allclose(axis.Z_n, eq.Za_n)
 
-        surface2 = ZernikeRZToroidalSection(spectral_indexing="ansi")
+        surface2 = FourierRZToroidalSurface(NFP=3)
         eq2 = Equilibrium(surface=surface2)
+        assert eq2.NFP == 3
+        assert eq2.axis.NFP == 3
         assert eq2.surface.equiv(surface2)
 
-        surface3 = FourierRZToroidalSurface(NFP=3)
-        eq3 = Equilibrium(surface=surface3)
-        assert eq3.NFP == 3
-        assert eq3.axis.NFP == 3
-
-        eq4 = Equilibrium(surface=surface2, axis=None)
-        np.testing.assert_allclose(eq4.axis.R_n, [10])
+        eq3 = Equilibrium(xsection=xsection, axis=None)
+        np.testing.assert_allclose(eq3.axis.R_n, [10])
+        assert eq3.xsection.equiv(xsection)
 
     @pytest.mark.unit
     def test_dict(self):
@@ -95,6 +94,7 @@ class TestConstructor:
                 ]
             ),
             "axis": np.array([[0, 10, 0]]),
+            "xsection": None,
             "pressure": np.array([[0, 10], [2, 5]]),
             "iota": np.array([[0, 1], [2, 3]]),
         }
@@ -169,22 +169,17 @@ class TestConstructor:
             ],
         )
 
-        inputs["surface"] = np.array(
+        inputs["xsection"] = np.array(
             [
                 [0, 0, 0, 10, 0],
                 [1, 1, 0, 1, 0.1],
                 [1, -1, 0, 0.2, -1],
             ]
         )
+        inputs.pop("surface")
 
-        eq = Equilibrium(**inputs)
-        assert eq.bdry_mode == "poincare"
-        np.testing.assert_allclose(
-            eq.Rb_lmn, [10.0, 0.2, 1.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-        )
-        np.testing.assert_allclose(
-            eq.Zb_lmn, [0.0, -1.0, 0.1, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]
-        )
+        with pytest.raises(NotImplementedError):
+            eq = Equilibrium(**inputs)
 
     @pytest.mark.unit
     def test_asserts(self):
@@ -272,10 +267,6 @@ class TestInitialGuess:
             eq.surface = eq.get_surface_at(rho=1)
             eq.change_resolution(2, 2, 2)
             _ = _initial_guess_surface(eq.R_basis, eq.R_lmn, eq.R_basis)
-        with pytest.raises(ValueError):
-            _ = _initial_guess_surface(
-                eq.R_basis, eq.surface.R_lmn, eq.surface.R_basis, mode="foo"
-            )
 
     @pytest.mark.unit
     def test_guess_from_other(self):
@@ -468,8 +459,14 @@ class TestGetSurfaces:
     def test_get_zeta_surface(self):
         """Test getting a constant zeta surface."""
         eq = Equilibrium()
-        surf = eq.get_surface_at(zeta=np.pi)
-        assert surf.zeta == np.pi
+        with pytest.raises(UserWarning):
+            surf = eq.get_surface_at(zeta=np.pi)
+            assert surf.zeta == np.pi
+            rho = 1
+            np.testing.assert_allclose(surf.compute("A")["A"], np.pi * rho**2)
+
+        surf = eq.get_surface_at(zeta=0)
+        assert surf.zeta == 0
         rho = 1
         np.testing.assert_allclose(surf.compute("A")["A"], np.pi * rho**2)